refactor(auth): move request structures to `model.go`
